Siirrä sure-kyselyiden jaettava koodi traittiin
